modified dice_roller__init__.py, test_integration.py, test_dice_roller, changed name of test_experiment_file.py to find_roles.py, moved this file to typo_hunting and modified this file
This is messy.  So the first three files I modified, I justed added a comment at the end which had role in it somewhere to help me test find_role.  find_role.py got overhauled and now should be working.
