+++ dice_roller/__init__.py
@@ -114,3 +114,5 @@ class DiceRoller:
 
         self._records = []
         return "History Cleared!"
+
+# TODO: DELETE THIS LINE LATER; TESTING 'role' TYPO HUNTING
