+++ tests/integration/test_integration.py
@@ -71,3 +71,5 @@ class DiceRollerIntegrationTests(unittest.TestCase):
         dice_roller_instance.clear()
 
         self.assertEqual(dice_roller_instance.history(), {})
+
+# TODO: DELETE THIS LINE LATER; TESTING 'ROLE' TYPO HUNTING rOLe
